+++ lib/concierge/roomorama/calendar/stay.rb
@@ -5,7 +5,6 @@ class Roomorama::Calendar::Stay
   attribute :checkin,   presence: true, type: Date, format: DATE_FORMAT
   attribute :checkout,  presence: true, type: Date, format: DATE_FORMAT
   attribute :price,     presence:true,  type: Float
-  attribute :available, presence:true
 
   def length
     (checkout - checkin).to_i