+++ CHANGELOG.md
@@ -11,6 +11,13 @@ how this file is formatted and how the process works.
 ### Added
 - Support for owner information when creating a `Roomorama::Property`
 
+### Fixed
+- `Roomorama::Calendar::StaysMapper` maps valid entries from tomorrow
+- `Roomorama::Calendar::StaysMapper` for empty stays case
+
+### Changed
+- Remove `Roomorama::Calendar::Stay.available` field
+
 ## [0.7.1] - 2016-08-17
 ### Changed
 - Better error handling for 404 and 500 scenarios on Roomorama webhook